refactor: rename test page
